Refactor ConditionBrInsn class
