+++ compiler/BIR.h
@@ -209,24 +209,6 @@ public:
   void translate(LLVMModuleRef &modRef) final;
 };
 
-class ConditionBrInsn : public TerminatorInsn {
-private:
-  BIRBasicBlock *ifThenBB;
-  BIRBasicBlock *elseBB;
-
-public:
-  ConditionBrInsn() = default;
-  ConditionBrInsn(Location *pos, InstructionKind kind, Operand *lOp,
-                  BIRBasicBlock *nextB);
-  ~ConditionBrInsn() = default;
-  void setIfThenBB(BIRBasicBlock *ifBB) { ifThenBB = ifBB; }
-  void setElseBB(BIRBasicBlock *elseB) { elseBB = elseB; }
-
-  BIRBasicBlock *getIfThenBB() { return ifThenBB; }
-  BIRBasicBlock *getElseBB() { return elseBB; }
-
-  void translate(LLVMModuleRef &modRef) final;
-};
