+++ compiler/codegen/BalConditionBrInsn.cpp
@@ -1,4 +1,18 @@
-#include "BIR.h"
+#include "BalBasicBlock.h"
+#include "BalConditionBrInsn.h"
+#include "BalFunction.h"
+#include "BalOperand.h"
+#include "BalTypeDecl.h"
+#include "BalVarDecl.h"
+#include "llvm-c/Core.h"
+
+#ifndef unix
+#define __attribute__(unused)
+#endif
+
+using namespace std;
+
+namespace nballerina {
 
 ConditionBrInsn::ConditionBrInsn(Location *pos, InstructionKind kind,
                                  Operand *lOp, BIRBasicBlock *nextBB)
@@ -33,3 +47,5 @@ void ConditionBrInsn::translate(__attribute__((unused)) LLVMModuleRef &modRef) {
     LLVMBuildCondBr(builder, brCondition, ifLLVMBB, elseLLVMBB);
   }
 }
+
+} // namespace nballerina
